python/requirements.txt
aiohttp>=3.9.5
discord.py>=2.3.2
python-dotenv>=1.0.1
requests>=2.31.0
